requirements.txt
textract
google-api-python-client
oauth2client
pyyaml
